% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjNetScore.R
\name{AdjNetScore}
\alias{AdjNetScore}
\title{AdjNetScore}
\usage{
AdjNetScore(TotalScore, NumSeedGene, NumScoreGene, multiple = 1000)
}
\arguments{
\item{TotalScore}{Total Score}

\item{NumSeedGene}{Number of Seed Gene}

\item{NumScoreGene}{Seed Gene Score}

\item{multiple}{Adjusted multiple constant}
}
\value{
a numeric object
}
\description{
Adjusted Network based relevance score
Adjusted based on the number of genes: Score=TotalScore/NumSeedGene*log10(1+NumSeedGene/NumScoreGene/2)*multiple
}
\author{
Yang Ming
}
