% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alertCal.R
\name{alertCal}
\alias{alertCal}
\title{alertCal}
\usage{
alertCal(sdfFile, unwanted_structureFile = "db/unwanted_structure.csv")
}
\arguments{
\item{sdfFile}{molecular file *.sdf}

\item{unwanted_structureFile}{a csv file with colnames:Name, smarts}
}
\value{
a dataframe object
}
\description{
Calculate ALERTS score in QED
}
\author{
Yang Ming
}
