% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asSymmetric.R
\name{asSymmetric}
\alias{asSymmetric}
\title{asSymmetric}
\usage{
asSymmetric(x, rule = "upper")
}
\arguments{
\item{x}{a matrix}
}
\value{
a matrix object
}
\description{
Make the matrix Symmetric
}
\author{
Yang Ming
}
