% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BF_analysis.R
\name{BF_analysis}
\alias{BF_analysis}
\title{BF_analysis}
\usage{
BF_analysis(
  Pdata,
  tS0.9 = 0.03,
  min_yao = 8,
  threshold_leastYao = 10,
  exact = FALSE,
  combineSim = FALSE,
  combine_method = 2,
  combine_threshold = 0.7,
  csvInput = T
)
}
\arguments{
\item{Pdata}{matrix,colnames=yao,rownames=prescription$id OR csv\link{colname:id,herb}}

\item{tS0.9}{S0.9 value}

\item{min_yao}{minimum number of herbs}

\item{threshold_leastYao}{expected number of herbs}

\item{exact}{combine methods}

\item{combineSim}{whether to combine similar Formulas}

\item{combine_method}{combine method to choose}

\item{combine_threshold}{threshold to combine}

\item{csvInput}{whether Pdata is a csv file}
}
\value{
a list Object
}
\description{
Prescription Mining
}
\author{
Yang Ming
}
