% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chem2DSim.R
\name{Chem2DSim}
\alias{Chem2DSim}
\title{Chem2DSim}
\usage{
Chem2DSim(SMI)
}
\arguments{
\item{SMI}{molecular file: a dataframe with smiles}
}
\value{
a list object
}
\description{
Calculate Compound 2D Similarity
}
\author{
Yang Ming
}
