% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chem3DSim.R
\name{Chem3DSim}
\alias{Chem3DSim}
\title{Chem3DSim}
\usage{
Chem3DSim(SMI)
}
\arguments{
\item{SMI}{molecular file: a dataframe with smiles}
}
\value{
a list object
}
\description{
Calculate Compound 3D Similarity
}
\author{
Yang Ming
}
