% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemScoreCal.R
\name{chemScoreCal}
\alias{chemScoreCal}
\title{chemScoreCal}
\usage{
chemScoreCal(chem_target, geneScore, selectGeneId, method = 1, cutoffValue = 0)
}
\arguments{
\item{chem_target}{data.frame-->colnames(chem_target)=c('id','geneID')}

\item{geneScore}{gene Score}

\item{selectGeneId}{Gene Id}

\item{method}{1: Use the cutoffValue; 2: SelectGeneId with a minimum of 100\% coverage.}

\item{cutoffValue}{cutoff Value}
}
\value{
a list Object
}
\description{
Compound Score
}
\author{
Yang Ming
}
