% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineIgraphOB.R
\name{combineIgraphOB}
\alias{combineIgraphOB}
\title{combineIgraphOB}
\usage{
combineIgraphOB(g1, g2)
}
\arguments{
\item{g1}{a igraph Object}

\item{g2}{a igraph Object}
}
\value{
a igraph object
}
\description{
Combine igraph Object
}
\author{
Yang Ming
}
