% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisAccess.R
\name{DisAccess}
\alias{DisAccess}
\title{DisAccess}
\usage{
DisAccess(SeedGeneID, ScoreGeneID, randomtimes = 1000, testM = 1)
}
\arguments{
\item{SeedGeneID}{geneID for Seed}

\item{ScoreGeneID}{geneID for Scoring}

\item{randomtimes}{randomtimes for permutaion test}

\item{testM}{test method:1-->wilcox;testM:2-->permutation}
}
\value{
a list Object
}
\description{
Network based relevance inference by network distance
}
\author{
Yang Ming
}
