% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichMentAnalysis.R
\name{EnrichMentAnalysis}
\alias{EnrichMentAnalysis}
\title{EnrichMentAnalysis}
\usage{
EnrichMentAnalysis(
  geneClusterID,
  type,
  qvalueCutoff = 0.05,
  filtKEGGdisease = F,
  ontGO = "BP",
  ontDO = "DOLite",
  ClusterNameToMESHenrich = names(geneClusterID)[1],
  meshDatabase = "gene2pubmed",
  meshcategory = "C"
)
}
\arguments{
\item{geneClusterID}{a named list}

\item{type}{Enrichment type:'KEGG' or 'GO' 'DO' or 'MESH' or 'ReactomePA'}

\item{qvalueCutoff}{Adjusted P value}

\item{filtKEGGdisease}{Remove KEGG disease terms}

\item{ontGO}{Enrichment GO type}

\item{ontDO}{Enrichment DO}

\item{ClusterNameToMESHenrich}{a named list for gene cluster}

\item{meshDatabase}{for MESH analysis}
}
\value{
a list Object
}
\description{
Enrichment Analysis
}
\author{
Yang Ming
}
