% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScoreCal.R
\name{geneScoreCal}
\alias{geneScoreCal}
\title{geneScoreCal}
\usage{
geneScoreCal(chem_target, Pcutoff)
}
\arguments{
\item{chem_target}{data.frame-->colnames(chem_target)=c('id','geneID')}

\item{Pcutoff}{Pvalue cutoff}
}
\value{
a list Object
}
\description{
Gene Score
}
\author{
Yang Ming
}
