% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPDBbox.R
\name{getPDBbox}
\alias{getPDBbox}
\title{getBoxFromFpocket}
\usage{
getPDBbox(
  PDBid,
  type = "FromLigand",
  center_x = NULL,
  center_y = NULL,
  center_z = NULL,
  size_x = NULL,
  size_y = NULL,
  size_z = NULL,
  python2PTAH = "~/mgltools_x86_64Linux2_1.5.6/bin/pythonsh",
  prepare_ligand4PATH = "~/ADtools",
  PDBboxDb = "PDBboxDB.db"
)
}
\arguments{
\item{PDBid}{PDB ID}

\item{type}{FromLigand;FromFpocket;Input}

\item{PDBboxDb}{prepared pocket parameter database}
}
\value{
a dataframe Object
}
\description{
Get protein docking pocket parameters.
Note: Fpocket and MGLTools must be installed and configured correctly.
}
\author{
Yang Ming
}
