% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCorr.R
\name{NetCorr}
\alias{NetCorr}
\title{NetCorr}
\usage{
NetCorr(
  SeedGeneID,
  ScoreGeneID,
  Methodnet,
  randomtimes = 1000,
  gamma = 0.7,
  testM = 1
)
}
\arguments{
\item{SeedGeneID}{geneID for Seed}

\item{ScoreGeneID}{geneID for Scoring}

\item{Methodnet}{method for Scoring:RW, KATZ or DIS}

\item{randomtimes}{randomtimes for permutaion test}

\item{gamma}{RW parameter}

\item{testM}{test method:1-->wilcox;testM:2-->permutation}
}
\value{
a list Object
}
\description{
Network based relevance inference
}
\author{
Yang Ming
}
