% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareProtein.R
\name{prepareProtein}
\alias{prepareProtein}
\title{prepareProtein}
\usage{
prepareProtein(
  pdbFile,
  pdbtype = "id",
  preparedProfile = "outPro.pdbqt",
  python2PTAH = "c:/Progra~2/MGLTools-1.5.6rc3/python2",
  prepare_receptor4PATH = system.file("tools", package = "tcmnpas"),
  getLigand = T,
  RefLigandFile = "Ligand",
  IFProLigandPrepared = T,
  repair = T
)
}
\arguments{
\item{pdbFile}{Receptor file}

\item{pdbtype}{Receptor type: id pdb or pdbqt}

\item{preparedProfile}{prepared file name for output: *.pdbqt}

\item{python2PTAH}{python2.exe PTAH}

\item{prepare_receptor4PATH}{"prepare_receptor4.py" PTAH. The default Vina receptor protein preparation program (addh–>true for adding hydrogen; nphs–>true for merging charges and removing non-polar hydrogen; lps–>true for merging charges and removing lone pairs; waters–>true for removing water molecules; nonstdres–>true for removing chains composed of less than 20 standard amino acid residues) is used.}

\item{getLigand}{whether to extract ligands from the protein}

\item{RefLigandFile}{If getLigand is True. RefLigandFile is the file name for output}

\item{IFProLigandPrepared}{whether to Prepare ligand}

\item{repair}{whether to repair Receptor by "-A hydrogens"}
}
\value{
a prepared receptor file
}
\description{
PrepareReceptor for Molecular Docking.
Note: vina, PSOvina,and mgltools must be installed and configured correctly.
}
\author{
Yang Ming
}
