% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QEDcal.R
\name{QEDcal}
\alias{QEDcal}
\title{QEDcal}
\usage{
QEDcal(druglikeDesData, QED_par = "db/QED_par.csv")
}
\arguments{
\item{druglikeDesData}{a dataframe with colname: c('id',colnames(QED_par))}

\item{QED_par}{QED calculation parameter from reference: a dataframe with colname(MW,ALOGP,HBA,HBD,PSA,ROTB,AROM,ALERTS}
}
\value{
a dataframe
}
\description{
QED of chemical calculation
}
\author{
Yang Ming
}
