% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryHerbChem.R
\name{queryHerbChem}
\alias{queryHerbChem}
\title{queryHerbChem}
\usage{
queryHerbChem(
  herbList = NULL,
  Qinchikey = NULL,
  database = c("HIT", "TCMSP", "TCMID"),
  HerbChemDB = "../npa/db/herb_ChemDataAll.db"
)
}
\arguments{
\item{herbList}{a character vector}

\item{Qinchikey}{inchikey of chemical for query}

\item{database}{a character vector:'HIT','TCMSP','TCMID'}

\item{HerbChemDB}{a prepared herb-chemical database file}
}
\value{
a dataframe
}
\description{
query Herb Chemical from database
}
\author{
Yang Ming
}
