% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDFtoMOL2.R
\name{SDFtoMOL2}
\alias{SDFtoMOL2}
\title{SDFtoMOL2}
\usage{
SDFtoMOL2(molFile, molName = "mole.mol2")
}
\arguments{
\item{molFile}{molecular file *.sdf}

\item{molName}{molecular file name}
}
\value{
a mol2 file
}
\description{
SDF to MOL2
note:obabel must be installed
}
\author{
Yang Ming
}
