% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDFtoSMILES.R
\name{SDFtoSMILES}
\alias{SDFtoSMILES}
\title{SDFtoSMILES}
\usage{
SDFtoSMILES(SDFfile, outtype = "txt", outName = "smiles.csv")
}
\arguments{
\item{SDFfile}{molecular file *.sdf}

\item{outtype}{'txt' or 'file'}

\item{outName}{a csv file with transformed SMILES character}
}
\value{
a character vector or a dataframe
}
\description{
SDF to SMILES
note:obabel must be installed
}
\author{
Yang Ming
}
