% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMItoMOL2.R
\name{SMItoMOL2}
\alias{SMItoMOL2}
\title{SMItoMOL2}
\usage{
SMItoMOL2(moltxt, molName = "mole.mol2")
}
\arguments{
\item{moltxt}{SMILES}

\item{molName}{output file name}
}
\value{
a mol2 file
}
\description{
SMILES to MOL2
note:obabel must be installed
}
\author{
Yang Ming
}
