% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmiToSDF.R
\name{SmiToSDF}
\alias{SmiToSDF}
\title{SmiToSDF}
\usage{
SmiToSDF(SMI, type = "C")
}
\arguments{
\item{SMI}{SMILES}

\item{type}{type='F'-->file for SMI or 'C'-->character for SMItext}
}
\value{
a sdf file
}
\description{
SMILES to SDF
note:obabel must be installed
}
\author{
Yang Ming
}
