% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCMNP.R
\name{TCMNP}
\alias{TCMNP}
\title{TCMNP}
\usage{
TCMNP(
  CustomherbTargetDat = NULL,
  herbList,
  herbListName = "TCM",
  QEDset = 0.2,
  geneScore = 400,
  targetDatabase = c("HIT", "TCMID", "STITCH", "TCMSP", "CUSTOM"),
  diseaseGeneID = NULL,
  geneSelectPv = 0.05,
  GOMFenrich = T,
  GOBPenrich = T,
  GOCCenrich = T,
  KEGGenrich = T,
  PAenrich = T,
  DOenrich = T,
  MESHenrich = T,
  meshcategory = "C",
  qvalueCutoff = 0.05,
  TopNIDCurve = 10,
  AllHerbListData = AllHerbListData,
  TarDB = "db/Tar.db"
)
}
\arguments{
\item{CustomherbTargetDat}{a dataframe with colname:herb,cid,chemical_name,geneID,score,inchikey,smiles,database,QED_DED}

\item{herbList}{a character vector}

\item{herbListName}{character}

\item{QEDset}{QED threshold}

\item{geneScore}{gene Score threshold}

\item{targetDatabase}{'HIT','TCMID','STITCH','TCMSP', or 'CUSTOM'}

\item{diseaseGeneID}{disease GeneID}

\item{GOMFenrich}{True or False}

\item{GOBPenrich}{True or False}

\item{GOCCenrich}{True or False}

\item{KEGGenrich}{True or False}

\item{PAenrich}{True or False}

\item{DOenrich}{True or False}

\item{meshcategory}{mesh enrichment category}

\item{qvalueCutoff}{enrichment adjusted P value}

\item{TopNIDCurve}{The association curve of the top N items shared by diseases and the compound}
}
\value{
a list Object
}
\description{
Formula Mechanism
}
\author{
Yang Ming
}
