\name{tcmnpas-package}
\alias{tcmnpas-package}
\alias{tcmnpas}
\docType{package}
\title{\packageTitle{tcmnpas}}
\description{\packageDescription{tcmnpas}}
\section{Package Content}{\packageIndices{tcmnpas}}
\author{\packageAuthor{tcmnpas}}
\section{Maintainer}{\packageMaintainer{tcmnpas}}
\keyword{package}
