% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinaDock.R
\name{vinaDock}
\alias{vinaDock}
\title{vinaDock}
\usage{
vinaDock(
  preparedLigand,
  preparedProtein,
  configfile = "966c_BOXconfig.txt",
  configfileTYPE = "file",
  center_x = NULL,
  center_y = NULL,
  center_z = NULL,
  size_x = NULL,
  size_y = NULL,
  size_z = NULL,
  exhaustiveness = 8,
  num_modes = 9,
  energy_range = 3,
  cpu = 2,
  outPDBQT = "DockedPDBQT.pdbqt",
  PSO = FALSE
)
}
\arguments{
\item{preparedLigand}{prepared Ligand}

\item{preparedProtein}{prepared Protein}

\item{configfile}{a txt file}

\item{configfileTYPE}{"file" OR "input"}

\item{center_x}{input parameter}

\item{center_y}{input parameter}

\item{center_z}{input parameter}

\item{size_x}{input parameter}

\item{size_y}{input parameter}

\item{size_z}{input parameter}

\item{exhaustiveness}{input parameter}

\item{num_modes}{input parameter}

\item{energy_range}{input parameter}

\item{cpu}{input parameter}

\item{outPDBQT}{output docked file name}

\item{PSO}{whether to use PSOvina}
}
\value{
a list Object
}
\description{
Molecular Docking.
Note: vina, PSOvina,and mgltools must be installed and configured correctly.
}
\author{
Yang Ming
}
